pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "deepresearch"
version = "0.1.0"
description = "LangGraph and LangChain deep research agent scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "deepagents>=0.0.7,<0.1.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "langchain>=0.3.0,<0.4.0",
  "langchain-openai>=0.2.0,<0.3.0",
  "langgraph>=0.3.0,<0.4.0",
  "langgraph-checkpoint-sqlite>=2.0.0,<3.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "tavily-python>=0.5.0,<0.6.0",
  "tenacity>=9.0.0,<10.0.0",
  "trafilatura>=1.9.0,<2.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

